requirements-dev.txt
pytest~=7.1.3
pytest-benchmark~=3.4.1
pytest-cov~=4.0.0
pytest-recording~=0.12.1
pytest-console-scripts~=1.3.1
recommonmark~=0.7.1
requests-mock~=1.10.0
Sphinx~=3.5.1

mypy~=0.990
types-requests~=2.28.11
types-python-dateutil~=2.8.19
flake8~=5.0.4
black~=22.10.0
codespell~=2.2.2

jsonschema~=4.7.2
coverage~=6.5
doc8~=1.0.0

pre-commit==2.20.0

# optional dependencies
orjson==3.8.1